cleancode修改
cleancode修改


# message auto-generated for no-merge-commit merge:
merge 202506/cleancode-0618 into master

Cleancode

Created-by: insightDeveloper

Match-id-ed1e04519e4f333dee85923f2230c7e5e0aa0847

+++ server/src/modules/source/parser/SourceInstructionParser.cpp
@@ -674,7 +674,7 @@ std::vector<SourceFileLineDynamicCol> SourceInstructionParser::GetApiLinesDynami
     if (targetCore == apiCores.end()) {
         targetCore = apiCores.begin();
     }
-    size_t index = std::distance(apiCores.begin(), targetCore); // never below zero
+    size_t index = static_cast<size_t>(std::distance(apiCores.begin(), targetCore)); // never below zero
     for (const auto &item: sourceLinesMap[sourceName]) {
         SourceFileLineDynamicCol col;
         GetValueInTargetCore(item.stringColumnMap, col.stringColumnMap, index);
